refactor: move test-analysis types to homeboy-extension-contract
Opens the test phase-subsystem breakdown with the same data/behavior split
proven on bench/trace. The five test-analysis types (TestFailure,
TestAnalysisInput, FailureCluster, FailureCategory, TestAnalysis) are a pure
self-contained closure - zero dependencies, only referencing each other.

The 11 analysis functions in test/analyze.rs (failure clustering/categorization
behavior) stay in core. Core re-exports the types so all crate::extension::test
paths stay stable.

Clears the TestFailure core-import edge and sets up further test/lint cuts
(TestCounts, TestCommandOutput, LintCommandOutput) toward severing the test/lint
phase behavior surface.
